:hammer: refact: Refatoração do código da classe Biblioteca para classes services
