--- AplicacaoBrasileiraBibliotecaAcademica/src/aplicacaobrasileirabibliotecaacademica/Operador.java
@@ -1,13 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
-package aplicacaobrasileirabibliotecaacademica;
-
-
-public interface Operador {
-	public abstract void criar();
-	public abstract void listar();
-	public abstract void editar();
-	public abstract void excluir();
-}
