+++ AplicacaoBrasileiraBibliotecaAcademica/src/biblioteca/telas/MenuPrincipal.java
@@ -1,12 +1,9 @@
 package biblioteca.telas;
 
-import biblioteca.model.Biblioteca;
-
 import javax.swing.*;
 
 public class MenuPrincipal {
 
-    protected Biblioteca biblioteca = new Biblioteca();
 
     private String msg ="Menu Biblioteca\n Escolha uma das opções abaixo: \n 1 - Livros\n 2 - Cadastros\n" +
                 " 3 - Emprestimos\n";
